# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a endpoint's example",
        "title": "Music Example API",
        "termsOfService": "italo.me",
        "contact": {
            "name": "Italo Esdrass",
            "email": "suport@italo.me"
        },
        "license": {
            "name": "Italo 1.0",
            "url": "italoesdrass.me"
        },
        "version": "0.1"
    },
    "host": "localhost:8000/api/v1",
    "basePath": "/",
    "paths": {
        "/music/create": {
            "post": {
                "description": "Create a new user in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "music"
                ],
                "summary": "Create a new user",
                "parameters": [
                    {
                        "description": "Email",
                        "name": "email",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Name",
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Date",
                        "name": "date",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "City",
                        "name": "city",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Contact",
                        "name": "contact",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/dto.BaseFailure"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.BaseFailure": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": false
                }
            }
        },
        "dto.BaseSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean",
                    "default": true
                }
            }
        }
    }
}